requirements.txt
torch==1.4.0
torchvision==0.5.0
evalutils==0.2.4
scikit-learn==0.23.2
scipy==1.5.2
scikit-image==0.17.2
pathlib
numpy==1.19.2
SimpleITK
nibabel
medpy
matplotlib==3.3.2
pillow==8.3.2
